Split settings runtime surface tests
