+++ docs/DECISIONS.md
@@ -266,7 +266,7 @@
 - Tool-result final-answer copy lives in `AgentFinalAnswerBuilder`, not in `AgentRunner`. `AgentRunner.finalAnswer(...)` remains the compatibility entry point, but shell/browser/file/patch/MCP/Computer Use response wording belongs in the focused builder with direct tests and parity gates.
 - Deterministic mock model behavior lives in `MockLLMClient.swift`, not in `Agent.swift`. The mock client can keep local smoke-test and harness heuristics, but the main agent file should stay focused on agent contracts, streaming preview helpers, execution orchestration, and run result types.
 - Streaming action collection and partial assistant-preview parsing live in `AgentActionStreaming.swift`, not in `Agent.swift`. `AgentRunner` and `TrustedRouterLLMClient` should delegate to `AgentActionStreamCollector`, and draft assistant text should go through `AgentActionStreamPreview` so incomplete streaming JSON, raw stream accumulation, and duplicate preview suppression are handled in one tested helper.
-- Sidebar action hierarchy should stay sparse. The visible rail keeps only `New chat` as a primary action; Search, Command palette, Plugins, Automations, Terminal, Browser, Memories, and Activity live in the shared Tools menu driven by `QuillCodeSidebarCommandPresentation.utilityCommandIDs`. This keeps Codex-style navigation calm while preserving quick access and prevents native SwiftUI, static HTML, and Playwright from drifting on sidebar action order.
+- Sidebar action hierarchy should stay sparse while matching the Codex first-read rail. The visible primary rail keeps `New chat`, `Search`, `Plugins`, and `Automations` as primary actions; Command Palette, Terminal, Browser, Memories, and Activity live in the shared Tools menu driven by `QuillCodeSidebarCommandPresentation.utilityCommandGroups`. This preserves quick access to the Codex-like entry points while preventing native SwiftUI, static HTML, and Playwright from drifting on secondary tool order.
 - SSH Remote project execution should keep command routing, path validation, and command planning in separate seams. `WorkspaceRemoteProjectToolExecutor` owns only tool routing and SSH execution, `WorkspaceRemoteProjectPath` owns remote path normalization/artifact labels, and `WorkspaceRemoteGitToolRequestPlanner` owns Git/GitHub/worktree command construction plus URL/artifact intent. This keeps remote support testable as SSH, QuillCloud relay, and future remote transports expand.
 - Claude design critique on 2026-06-23 identified the next high-impact UI cleanup: keep the top bar quiet, move model selection and approval mode to the composer, simplify the sidebar into composer/threads/projects/footer zones, replace noisy bulk-action grids with transient horizontal actions, replace the flat Tools popover with a structured activity surface, teach Auto/Review/Read-only as first-class controls in the empty state, and make approvals a persistent inline queue banner instead of a modal-feeling interruption. The first slice moved model/mode into composer controls, hid the empty Chats header, balanced the Tools/Settings footer, and added native/static/Playwright coverage; the remaining items stay as UX roadmap inputs.
 - Tool-step execution lives in `AgentToolStepRunner.swift`, not in `Agent.swift`. `AgentRunner.send` owns the high-level model/tool loop and repeated-call fallback; the extracted runner owns safety review, tool lifecycle event emission, unavailable-tool results, apply-patch diff follow-up execution, and tool feedback serialization.
